+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * different bindings for `value={exp}` and `value="{exp}"`
 * deprecate function binding `@click={(e) => click(e)}`
 * improvements for !no-check
+* able to use `$watch(fn, callback, lvl)`
 
 ## 0.7.x
 * refactoring, optimization, fixes